server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ultralytics==8.3.0
python-multipart==0.0.17
pillow==11.0.0
numpy==2.1.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
